# Conflicts:
#	backend/src/main/java/org/ssafy/load/application/DeliveryService.java
#	backend/src/main/java/org/ssafy/load/common/dto/ErrorCode.java
#	backend/src/main/java/org/ssafy/load/dao/BuildingRepository.java
#	backend/src/main/java/org/ssafy/load/domain/AreaEntity.java
#	backend/src/main/java/org/ssafy/load/domain/BoxTypeEntity.java
#	backend/src/main/java/org/ssafy/load/dto/Building.java
#	backend/src/main/java/org/ssafy/load/dto/Goods.java

+++ backend/src/main/java/org/ssafy/load/domain/BoxTypeEntity.java
@@ -14,17 +14,21 @@ import org.ssafy.load.common.type.BoxType;
 @NoArgsConstructor
 @AllArgsConstructor
 public class BoxTypeEntity {
+
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private Integer id;
+
     @Enumerated(EnumType.STRING)
     private BoxType type;
+
     private int height;
     private int length;
     private int width;
 
+
     public static BoxTypeEntity of(
-            Long id,
+            Integer id,
             BoxType type,
             int height,
             int length,
